requirements.txt
flask==2.2.3
flask-sqlalchemy==3.0.0
flask-marshmallow==0.15.0
marshmallow==3.18.0
marshmallow-sqlalchemy==0.29.0
flask-jwt-extended==4.4.4
python-dotenv==1.0.0
werkzeug==2.2.3
pyjwt==2.6.0
sqlalchemy==1.4.46

Jinja2==3.1.6
weasyprint==65.0
qrcode==8.1
pillow==11.1.0
requests==2.31.0